Disable c_bindings feature by default
