+++ arkworks3/src/eip_7594.rs
@@ -2,8 +2,6 @@ extern crate alloc;
 
 use kzg::EcBackend;
 
-use kzg::c_bindings_eip7594;
-
 use crate::kzg_proofs::FFTSettings;
 use crate::kzg_proofs::KZGSettings;
 use crate::kzg_types::ArkFp;
@@ -26,4 +24,5 @@ impl EcBackend for ArkBackend {
     type KZGSettings = KZGSettings;
 }
 
-c_bindings_eip7594!(ArkBackend);
+#[cfg(feature = "c_bindings")]
+kzg::c_bindings_eip7594!(ArkBackend);
