+++ arkworks3/src/utils.rs
@@ -1,13 +1,11 @@
 extern crate alloc;
 
-use alloc::boxed::Box;
-
 use kzg::{
     eip_4844::PrecomputationTableManager,
     eth::{self, c_bindings::CKZGSettings},
 };
 
-use crate::kzg_proofs::{FFTSettings as LFFTSettings, KZGSettings as LKZGSettings};
+use crate::kzg_proofs::FFTSettings as LFFTSettings;
 use crate::kzg_types::{ArkFp, ArkFr, ArkG1, ArkG1Affine};
 
 use super::{Fp, P1};
@@ -27,17 +25,6 @@ pub struct PolyData {
     pub coeffs: Vec<ArkFr>,
 }
 // FIXME: Store just dense poly here
-
-macro_rules! handle_ckzg_badargs {
-    ($x: expr) => {
-        match $x {
-            Ok(value) => value,
-            Err(_) => return kzg::eth::c_bindings::CKzgRet::BadArgs,
-        }
-    };
-}
-
-pub(crate) use handle_ckzg_badargs;
 
 pub(crate) fn fft_settings_to_rust(
     c_settings: *const CKZGSettings,
@@ -89,88 +76,6 @@ pub(crate) static mut PRECOMPUTATION_TABLES: PrecomputationTableManager<
     ArkFp,
     ArkG1Affine,
 > = PrecomputationTableManager::new();
-
-pub(crate) fn kzg_settings_to_c(rust_settings: &LKZGSettings) -> CKZGSettings {
-    CKZGSettings {
-        roots_of_unity: Box::leak(
-            rust_settings
-                .fs
-                .roots_of_unity
-                .iter()
-                .map(|r| r.to_blst_fr())
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        brp_roots_of_unity: Box::leak(
-            rust_settings
-                .fs
-                .brp_roots_of_unity
-                .iter()
-                .map(|r| r.to_blst_fr())
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        reverse_roots_of_unity: Box::leak(
-            rust_settings
-                .fs
-                .reverse_roots_of_unity
-                .iter()
-                .map(|r| r.to_blst_fr())
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        g1_values_monomial: Box::leak(
-            rust_settings
-                .g1_values_monomial
-                .iter()
-                .map(|r| r.to_blst_p1())
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        g1_values_lagrange_brp: Box::leak(
-            rust_settings
-                .g1_values_lagrange_brp
-                .iter()
-                .map(|r| r.to_blst_p1())
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        g2_values_monomial: Box::leak(
-            rust_settings
-                .g2_values_monomial
-                .iter()
-                .map(|r| r.to_blst_p2())
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        x_ext_fft_columns: Box::leak(
-            rust_settings
-                .x_ext_fft_columns
-                .iter()
-                .map(|r| {
-                    Box::leak(
-                        r.iter()
-                            .map(|it| it.to_blst_p1())
-                            .collect::<Vec<_>>()
-                            .into_boxed_slice(),
-                    )
-                    .as_mut_ptr()
-                })
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        tables: core::ptr::null_mut(),
-        wbits: 0,
-        scratch_size: 0,
-    }
-}
 
 pub fn pc_poly_into_blst_poly(poly: DensePoly<ark_bls12_381::Fr>) -> PolyData {
     let mut bls_pol: Vec<ArkFr> = { Vec::new() };
