+++ blst/src/utils.rs
@@ -1,10 +1,8 @@
 extern crate alloc;
 
-use alloc::boxed::Box;
 use alloc::vec::Vec;
 
 use kzg::eip_4844::{hash_to_bls_field, PrecomputationTableManager};
-use kzg::eth::c_bindings::CKZGSettings;
 use kzg::{Fr, G1Mul, G2Mul};
 
 use crate::consts::{G1_GENERATOR, G2_GENERATOR};
@@ -12,7 +10,6 @@ use crate::types::fp::FsFp;
 use crate::types::fr::FsFr;
 use crate::types::g1::{FsG1, FsG1Affine};
 use crate::types::g2::FsG2;
-use crate::types::kzg_settings::FsKZGSettings;
 
 pub fn generate_trusted_setup(
     n: usize,
@@ -36,102 +33,9 @@ pub fn generate_trusted_setup(
     (s1, s2, s3)
 }
 
-macro_rules! handle_ckzg_badargs {
-    ($x: expr) => {
-        match $x {
-            Ok(value) => value,
-            Err(_) => return kzg::eth::c_bindings::CKzgRet::BadArgs,
-        }
-    };
-}
-
-pub(crate) use handle_ckzg_badargs;
-
 pub(crate) static mut PRECOMPUTATION_TABLES: PrecomputationTableManager<
     FsFr,
     FsG1,
     FsFp,
     FsG1Affine,
 > = PrecomputationTableManager::new();
-
-pub(crate) fn kzg_settings_to_c(rust_settings: &FsKZGSettings) -> CKZGSettings {
-    CKZGSettings {
-        roots_of_unity: Box::leak(
-            rust_settings
-                .fs
-                .roots_of_unity
-                .iter()
-                .map(|r| r.0)
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        brp_roots_of_unity: Box::leak(
-            rust_settings
-                .fs
-                .brp_roots_of_unity
-                .iter()
-                .map(|r| r.0)
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        reverse_roots_of_unity: Box::leak(
-            rust_settings
-                .fs
-                .reverse_roots_of_unity
-                .iter()
-                .map(|r| r.0)
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        g1_values_monomial: Box::leak(
-            rust_settings
-                .g1_values_monomial
-                .iter()
-                .map(|r| r.0)
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        g1_values_lagrange_brp: Box::leak(
-            rust_settings
-                .g1_values_lagrange_brp
-                .iter()
-                .map(|r| r.0)
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        g2_values_monomial: Box::leak(
-            rust_settings
-                .g2_values_monomial
-                .iter()
-                .map(|r| r.0)
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        x_ext_fft_columns: Box::leak(
-            rust_settings
-                .x_ext_fft_columns
-                .iter()
-                .map(|r| {
-                    Box::leak(
-                        r.iter()
-                            .map(|it| it.0)
-                            .collect::<Vec<_>>()
-                            .into_boxed_slice(),
-                    )
-                    .as_mut_ptr()
-                })
-                .collect::<Vec<_>>()
-                .into_boxed_slice(),
-        )
-        .as_mut_ptr(),
-        tables: core::ptr::null_mut(),
-        wbits: 0,
-        scratch_size: 0,
-    }
-}
