+++ constantine/Cargo.toml
@@ -26,7 +26,6 @@ default = [
     "std",
     "rand",
     "bgmw",
-    "c_bindings"
 ]
 std = [
     "hex/std",
