+++ zkcrypto/Cargo.toml
@@ -24,7 +24,6 @@ rand = { version = "0.8.5" }
 default = [
     "std",
     "rand",
-    "c_bindings"
 ]
 std = [
     "kzg/std",
